# Conflicts:
#	controllers/deelnemers_controller.js

+++ controllers/deelnemers_controller.js
@@ -8,25 +8,40 @@ module.exports = {
         console.log('added particitant to the meal');
 
         const StudentenhuisID = req.params.id;
-        let userId = 1; // To be changed according to token values
-        //const UserID = req.body.UserID;
+        let UserID = 1; // To be changed according to token values
         const MaaltijdID = req.params.maaltijdId;
 
+        console.log(req.body);
+        console.log("HuisID: " + StudentenhuisID);
+        console.log("MaaltijdID: "+ MaaltijdID);
+
         try {
-            expect(UserID).to.be.at.least(0 ["Id munst be a number and above 0"]);
-            expect(StudentenhuisID).to.be.at.least(0 ["UserId munst be a number and above 0"]);
-            expect(MaaltijdID).to.be.at.least(0 ["UserId munst be a number and above 0"]);
-            expect(UserID).to.not.to.be.empty;
-            expect(StudentenhuisID).to.not.to.be.empty;
-            expect(MaaltijdID).to.not.to.be.empty;
-            expect(res.body).should.be.a('object');
+            // expect(UserID).to.be.at.least(0 ["Id munst be a number and above 0"]);
+            // expect(StudentenhuisID).to.be.at.least(0 ["UserId munst be a number and above 0"]);
+            // expect(MaaltijdID).to.be.at.least(0 ["UserId munst be a number and above 0"]);
+            // expect(UserID).to.not.to.be.empty;
+            // expect(StudentenhuisID).to.not.to.be.empty;
+            // expect(MaaltijdID).to.not.to.be.empty;
+            //expect([UserID, StudentenhuisID, MaaltijdID]).to.not.exist();
+            //expect(res.body).should.be.a('object');
         } catch (ex) {
             const error = new ApiError(ex.toString(), 422)
             next(error)
             return
         }
 
-        db.query('INSERT INTO `deelnemers` (UserID, StudentenhuisID, MaaltijdID) VALUES (' + UserID + ', ' + StudentenhuisID + ', ' + MaaltijdID + ')', (error, rows, fields) => {
+        db.query('SELECT UserID FROM deelnemers WHERE StudentenhuisID = ' + StudentenhuisID + ' AND MaaltijdID = ' + MaaltijdID ,(error, rows, fields) => {
+            if(error) {
+                next(error);
+            } else {
+                if(rows[0]) {
+                    //User exists
+                    res.status(409).json({"error": "already a participant"});
+                }
+            }
+        });
+
+        db.query('INSERT INTO `deelnemers` (UserID, StudentenhuisID, MaaltijdID) VALUES (' + UserID + ', ' + StudentenhuisID + ', ' + MaaltijdID + ');', (error, rows, fields) => {
             if (error) {
                 next(error);
             } else {
@@ -51,31 +66,49 @@ module.exports = {
     },
     deleteParticipant(req, res, next) {
         const StudentenhuisID = req.params.id;
-        let userId = 1; // To be changed according to token values
+        let UserID = 1; // To be changed according to token values
         const MaaltijdID = req.params.maaltijdId;
 
+        console.log(req.body);
+        console.log("HuisID: " + StudentenhuisID);
+        console.log("MaaltijdID: "+ MaaltijdID);
+
         try {
-            expect(UserID).to.be.at.least(0 ["Id munst be a number and above 0"]);
-            expect(StudentenhuisID).to.be.at.least(0 ["UserId munst be a number and above 0"]);
-            expect(MaaltijdID).to.be.at.least(0 ["UserId munst be a number and above 0"]);
-            expect(UserID).to.not.to.be.empty;
-            expect(StudentenhuisID).to.not.to.be.empty;
-            expect(MaaltijdID).to.not.to.be.empty;
-            expect(res.body).should.be.a('object');
+            // expect(UserID).to.be.at.least(0 ["Id munst be a number and above 0"]);
+            // expect(StudentenhuisID).to.be.at.least(0 ["UserId munst be a number and above 0"]);
+            // expect(MaaltijdID).to.be.at.least(0 ["UserId munst be a number and above 0"]);
+            // expect(UserID).to.not.to.be.empty;
+            // expect(StudentenhuisID).to.not.to.be.empty;
+            // expect(MaaltijdID).to.not.to.be.empty;
+            // expect(UserID, StudentenhuisID, MaaltijdID).to.exist();
+            //expect(res.body).should.be.a('object');
         } catch (ex) {
             const error = new ApiError(ex.toString(), 422)
             next(error)
             return
         }
 
-        db.query('DELETE FROM `deelnemers` WHERE UserID = ' + UserID + ' AND StudentenhuisID = ' + StudentenhuisID + ' AND MaaltijdID =' + MaaltijdID, (error, rows, fields) => {
-            if (error) {
+        db.query('SELECT UserID FROM deelnemers WHERE StudentenhuisID = ' + StudentenhuisID + ' AND MaaltijdID = ' + MaaltijdID ,(error, rows, fields) => {
+            if(error) {
                 next(error);
             } else {
-                console.log('You have deleted nr: ' + id);
-                res.status(200).end();
+                if(rows[0]) {
+                    //User exists
+                    db.query('DELETE FROM `deelnemers` WHERE UserID = ' + UserID + ' AND StudentenhuisID = ' + StudentenhuisID + ' AND MaaltijdID = ' + MaaltijdID, (error, rows, fields) => {
+                        if (error) {
+                            next(error);
+                        } else {
+                            console.log('You have deleted nr: ' + id);
+                            res.status(200).end();
+
+                        }
+                    });
+                }else{
+                    res.status(409).json({"error": "already deleted"});
+                }
             }
-            ;
         });
+
+
     }
 }
